feature/juan/clean Feat: Update the CharacterDetailViewModel
